primeiras validações referentes à existência de campos de email, nome e password.
